# blogproject/blogapi/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Blog Web APIs",
    "description": "A basic project to test APIs, JWT and Swagger",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Bearer token to access these api endpoints",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/comments/{blogID}": {
      "post": {
        "tags": ["Comments"],
        "summary": "Add comment to a blog",

        "parameters": [
          {
            "name": "blogID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "example": "comment here"
                },
                "author": {
                  "example": "author id here"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": ["Comments"],
        "summary": "Get all comments of a blog",
        "parameters": [
          {
            "name": "blogID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/comments/onecomment/{_id}": {
      "get": {
        "tags": ["Comments"],
        "summary": "Get a comment of its ID",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/comments/{blogID}/{commentID}": {
      "put": {
        "tags": ["Comments"],
        "summary": "Update a comment in a blog",
        "parameters": [
          {
            "name": "blogID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "commentID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "example": "content here"
                },
                "authordId": {
                  "example": "author id here"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["Comments"],
        "summary": "Delete a comment in a blog",
        "parameters": [
          {
            "name": "blogID",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "commentID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get user by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },

      "put": {
        "tags": ["User"],
        "summary": "Update a user by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "your name"
                },

                "email": {
                  "example": "@gmail.com"
                },
                "password": {
                  "example": "password"
                },
                "role": {
                  "example": "user"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Delete a user by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}